deal label migration and miner test type migration integrated into the same migration again
