+++ actors/migration/nv14/top.go
@@ -7,15 +7,20 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/filecoin-project/go-state-types/big"
+
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/rt"
 	builtin5 "github.com/filecoin-project/specs-actors/v5/actors/builtin"
+	market5 "github.com/filecoin-project/specs-actors/v5/actors/builtin/market"
+
 	states5 "github.com/filecoin-project/specs-actors/v5/actors/states"
 	builtin6 "github.com/filecoin-project/specs-actors/v6/actors/builtin"
+	"github.com/filecoin-project/specs-actors/v6/actors/builtin/market"
+	"github.com/filecoin-project/specs-actors/v6/actors/builtin/power"
 	states6 "github.com/filecoin-project/specs-actors/v6/actors/states"
-	adt6 "github.com/filecoin-project/specs-actors/v6/actors/util/adt"
-
+	"github.com/filecoin-project/specs-actors/v6/actors/util/adt"
 	"github.com/ipfs/go-cid"
 	cbor "github.com/ipfs/go-ipld-cbor"
 	"golang.org/x/sync/errgroup"
@@ -47,7 +52,7 @@ func ActorHeadKey(addr address.Address, head cid.Cid) string {
 	return addr.String() + "-h-" + head.String()
 }
 
-// Migrates from v13 to v14
+// Migrates from v14 to v15
 //
 // This migration only updates the actor code CIDs in the state tree.
 // MigrationCache stores and loads cached data. Its implementation must be threadsafe
@@ -55,6 +60,139 @@ type MigrationCache interface {
 	Write(key string, newCid cid.Cid) error
 	Read(key string) (bool, cid.Cid, error)
 	Load(key string, loadFunc func() (cid.Cid, error)) (cid.Cid, error)
+}
+
+func IsTestPostProofType(proofType abi.RegisteredPoStProof) bool {
+	testPoStProofTypes := [6]abi.RegisteredPoStProof{abi.RegisteredPoStProof_StackedDrgWinning2KiBV1,
+		abi.RegisteredPoStProof_StackedDrgWinning8MiBV1,
+		abi.RegisteredPoStProof_StackedDrgWinning512MiBV1,
+		abi.RegisteredPoStProof_StackedDrgWindow2KiBV1,
+		abi.RegisteredPoStProof_StackedDrgWindow8MiBV1,
+		abi.RegisteredPoStProof_StackedDrgWindow512MiBV1,
+	}
+	for i := 0; i < 6; i++ {
+		if proofType == testPoStProofTypes[i] {
+			return true
+		}
+	}
+	return false
+}
+
+// XXX balanceTransferMap should become a BalanceTable, maybe?
+func addIfPresent(balanceTransferMap *map[address.Address]big.Int, address address.Address, balance big.Int) {
+	if balance.Equals(big.Zero()) { // do nothing if it's equal...
+		return
+	}
+	if balance.LessThan(big.Zero()) {
+		return
+	} // XXX what are you doing why is it negative??? should we put an assertion here???
+
+	oldVal, found := (*balanceTransferMap)[address]
+	if !found {
+		(*balanceTransferMap)[address] = balance
+	} else {
+		(*balanceTransferMap)[address] = big.Add(balance, oldVal)
+	}
+}
+func MigrateMarketActor(ctx context.Context, store cbor.IpldStore, actorsIn *states5.Tree, actorsOut *states5.Tree) error {
+	marketActor, found, err := actorsIn.GetActor(builtin5.StorageMarketActorAddr)
+	if err != nil || !found {
+		return xerrors.Errorf("Could not get market actor for migration.")
+	}
+	var inState market5.State
+	if err := store.Get(ctx, marketActor.Head, &inState); err != nil {
+		return err
+	}
+	wrappedStore := adt.WrapStore(ctx, store)
+
+	proposalsCidOut, err := MapProposals(ctx, wrappedStore, inState.Proposals)
+	if err != nil {
+		return err
+	}
+
+	pendingProposalsCidOut, err := CreateNewPendingProposals(ctx, wrappedStore, proposalsCidOut, inState.States)
+	if err != nil {
+		return err
+	}
+
+	outState := market.State{
+		Proposals:                     proposalsCidOut,
+		States:                        inState.States,
+		PendingProposals:              pendingProposalsCidOut,
+		EscrowTable:                   inState.EscrowTable,
+		LockedTable:                   inState.LockedTable,
+		NextID:                        inState.NextID,
+		DealOpsByEpoch:                inState.DealOpsByEpoch,
+		LastCron:                      inState.LastCron,
+		TotalClientLockedCollateral:   inState.TotalClientLockedCollateral,
+		TotalProviderLockedCollateral: inState.TotalProviderLockedCollateral,
+		TotalClientStorageFee:         inState.TotalClientStorageFee,
+	}
+
+	newHead, err := store.Put(ctx, &outState)
+	if err != nil {
+		return nil
+	}
+
+	marketActor.Head = newHead
+
+	return actorsOut.SetActor(builtin6.StorageMarketActorAddr, marketActor)
+}
+
+// migrates the power actor for this actor type. deletes claims of invalid miner types, and updates MinerCount and MinerAboveMinPowerCount according to how they were computed before.
+func MigratePowerActor(ctx context.Context, store cbor.IpldStore, actorsIn *states5.Tree,
+	actorsOut *states5.Tree, totalDeletedMiners int, totalDeletedMinersAboveMinPower int,
+) error {
+	powerActor, found, err := actorsIn.GetActor(builtin5.StoragePowerActorAddr)
+	if err != nil || !found {
+		return xerrors.Errorf("Could not get power actor for migration.")
+	}
+	var powerState power.State
+	if err := store.Get(ctx, powerActor.Head, &powerState); err != nil {
+		return err
+	}
+	wrappedStore := adt.WrapStore(ctx, store)
+	claims, err := adt.AsMap(wrappedStore, powerState.Claims, builtin5.DefaultHamtBitwidth)
+	if err != nil {
+		return err
+	}
+
+	var claim power.Claim
+	err = claims.ForEach(&claim, func(key string) error {
+		if IsTestPostProofType(claim.WindowPoStProofType) {
+			// this just deletes any claims that ended up in the power actor's claims table which had the test miner types
+			// handling the power statistics (MinerCount, MinerAboveMinPowerCount) will be done in top.go after we've collected all that from the miners.
+			addr, err := address.NewFromString(key)
+			if err != nil {
+				return err
+			}
+			if claim.RawBytePower.GreaterThan(big.Zero()) {
+				return xerrors.Errorf("nonzero RawBytePower on claim from miner with test proof size. This is not good.")
+			}
+			if claim.QualityAdjPower.GreaterThan(big.Zero()) {
+				return xerrors.Errorf("nonzero QualityAdjPower on claim from miner with test proof size. This is not good.")
+			}
+			// XXX: I think you might need to manually change claims. this might be wrong. you need to write back a CID somewhere... do you?
+			if _, err := powerState.DeleteClaim(claims, addr); err != nil {
+				return err
+			}
+		}
+		return nil
+	})
+	if err != nil {
+		return nil
+	}
+	powerState.MinerCount -= int64(totalDeletedMiners)
+	powerState.MinerAboveMinPowerCount -= int64(totalDeletedMinersAboveMinPower)
+
+	newHead, err := store.Put(ctx, &powerState)
+	if err != nil {
+		return nil
+	}
+
+	powerActor.Head = newHead
+
+	return actorsOut.SetActor(builtin6.StoragePowerActorAddr, powerActor)
 }
 
 // Migrates the filecoin state tree starting from the global state tree and upgrading all actor state.
@@ -72,16 +210,17 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, actorsRootIn ci
 		builtin5.MultisigActorCodeID:         nilMigrator{builtin6.MultisigActorCodeID},
 		builtin5.PaymentChannelActorCodeID:   nilMigrator{builtin6.PaymentChannelActorCodeID},
 		builtin5.RewardActorCodeID:           nilMigrator{builtin6.RewardActorCodeID},
-		builtin5.StorageMarketActorCodeID:    cachedMigration(cache, marketMigrator{}),
-		builtin5.StorageMinerActorCodeID:     nilMigrator{builtin6.StorageMinerActorCodeID},
-		builtin5.StoragePowerActorCodeID:     nilMigrator{builtin6.StoragePowerActorCodeID},
+		builtin5.StorageMinerActorCodeID:     cachedMigration(cache, minerMigrator{}),
 		builtin5.SystemActorCodeID:           nilMigrator{builtin6.SystemActorCodeID},
 		builtin5.VerifiedRegistryActorCodeID: nilMigrator{builtin6.VerifiedRegistryActorCodeID},
 	}
 
 	// Set of prior version code CIDs for actors to defer during iteration, for explicit migration afterwards.
 	var deferredCodeIDs = map[cid.Cid]struct{}{
-		// None
+		builtin5.StoragePowerActorCodeID: {},
+		// "defer" market (do it first)
+		// this is so that we can migrate dealLabels earlier and then futz with all the escrow stuff while the other migrations are happening
+		builtin5.StorageMarketActorCodeID: {},
 	}
 
 	if len(migrations)+len(deferredCodeIDs) != 11 {
@@ -90,7 +229,7 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, actorsRootIn ci
 	startTime := time.Now()
 
 	// Load input and output state trees
-	adtStore := adt6.WrapStore(ctx, store)
+	adtStore := adt.WrapStore(ctx, store)
 	actorsIn, err := states5.LoadTree(adtStore, actorsRootIn)
 	if err != nil {
 		return cid.Undef, err
@@ -108,6 +247,10 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, actorsRootIn ci
 	// Atomically-modified counters for logging progress
 	var jobCount uint32
 	var doneCount uint32
+
+	if err := MigrateMarketActor(ctx, store, actorsIn, actorsOut); err != nil {
+		return cid.Undef, err
+	}
 
 	// Iterate all actors in old state root to create migration jobs for each non-deferred actor.
 	grp.Go(func() error {
@@ -200,22 +343,137 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, actorsRootIn ci
 		return nil
 	})
 
+	// building up a map of balance transfers- address to amount
+	// this mutex will only get held like 30 times over a map of all actors, so it will have zero contention, but better safe than sorry!
+	var balanceTransferMapGuard = &sync.Mutex{}
+	var balanceTransfers = make(map[address.Address]big.Int)
+	var totalDeletedMinersAboveMinPower = 0
+	var totalDeletedMiners = 0
+
+	marketActor, found, err := actorsIn.GetActor(builtin5.StorageMarketActorAddr)
+	if err != nil {
+		return cid.Undef, err
+	}
+	if !found {
+		return cid.Undef, fmt.Errorf("failed to get market actor for second migration")
+	}
+	var marketState market.State
+	if err := store.Get(ctx, marketActor.Head, &marketState); err != nil {
+		return cid.Undef, err
+	}
+	wrappedStore := adt.WrapStore(ctx, store)
+	escrowTable, err := adt.AsBalanceTable(wrappedStore, marketState.EscrowTable)
+	if err != nil {
+		return cid.Undef, err
+	}
+
 	// Insert migrated records in output state tree and accumulators.
+	// also, remove miners from tree if needed, and check if they're in the marketstate escrow table
+	// also remove them from the market state escrow table if needed
 	grp.Go(func() error {
 		log.Log(rt.INFO, "Result writer started")
 		resultCount := 0
+		deletedActorCount := 0
 		for result := range jobResultCh {
-			if err := actorsOut.SetActor(result.Address, &result.Actor); err != nil {
-				return err
+			if result.minerTypeMigrationShouldDelete {
+				balanceTransferMapGuard.Lock()
+				ownerAddressT := result.minerTypeMigrationBalanceTransferInfo.address
+				valueT := result.minerTypeMigrationBalanceTransferInfo.value
+				if !valueT.GreaterThanEqual(big.Zero()) {
+					return xerrors.Errorf("deleted test miner's balance was negative and we tried to send it to address %v", ownerAddressT)
+				}
+				// transfer miner's balance back to owner
+				addIfPresent(&balanceTransfers, ownerAddressT, valueT)
+
+				// prep to adjust power stats
+				if result.minerDeletedWasAboveMinPower {
+					totalDeletedMinersAboveMinPower++
+				}
+				totalDeletedMiners++
+				deletedActorCount++
+
+				// adjust miner escrow table and send funds around- first, get the amount
+				tokenAmountInEscrow, err := escrowTable.Get(result.address)
+				if err != nil {
+					return err
+				}
+				if !tokenAmountInEscrow.GreaterThanEqual(big.Zero()) {
+					return xerrors.Errorf("miner's balance in escrow was negative and we tried to send it to address %v", ownerAddressT)
+				}
+				// zero out miner balances from escrow table
+				if err = escrowTable.MustSubtract(result.address, tokenAmountInEscrow); err != nil {
+					return nil
+				}
+				// schedule that amount to go back to the owner
+				addIfPresent(&balanceTransfers, ownerAddressT, tokenAmountInEscrow)
+				balanceTransferMapGuard.Unlock()
+			} else {
+				if err := actorsOut.SetActor(result.address, &result.actor); err != nil {
+					return err
+				}
+				resultCount++
 			}
-			resultCount++
 		}
-		log.Log(rt.INFO, "Result writer wrote %d results to state tree after %v", resultCount, time.Since(startTime))
+		log.Log(rt.INFO, "Result writer wrote %d results to state tree and deleted %d actors after %v", resultCount, deletedActorCount, time.Since(startTime))
 		return nil
 	})
 
 	if err := grp.Wait(); err != nil {
 		return cid.Undef, err
+	}
+
+	// remove claims from these actors, reset power stats
+	if err = MigratePowerActor(ctx, store, actorsIn, actorsOut,
+		totalDeletedMiners, totalDeletedMinersAboveMinPower); err != nil {
+		return cid.Undef, err
+	}
+
+	// write back new escrow table into market state
+	newEscrowTableRoot, err := escrowTable.Root()
+	if err != nil {
+		return cid.Undef, err
+	}
+	marketState.EscrowTable = newEscrowTableRoot
+	marketActor.Head, err = store.Put(ctx, &marketState)
+	if err != nil {
+		return cid.Undef, err
+	}
+	if actorsOut.SetActor(builtin6.StorageMarketActorAddr, marketActor) != nil {
+		return cid.Undef, err
+	}
+
+	// doing balance increments for owners of the deleted miners with test state tree types
+	// this is balances from the miners' balances in the tree and also from the escrow tables in the market actor.
+
+	for addressT, valueT := range balanceTransfers { //nolint:nomaprange // this nolint is ok, this thing will be like 30 entries long max...
+		// check and make sure this is positive... just as a fun invariant, haha
+
+		incrementaddr := addressT
+		actor, found, err := actorsOut.GetActor(addressT)
+		if err != nil {
+			return cid.Undef, err
+		}
+		// if you don't find the owner of the deleted miner, swap to sending funds to f099
+		if !found {
+			f099addr, err := address.NewFromString("f099")
+			if err != nil {
+				return cid.Undef, err
+			}
+			actor, found, err = actorsOut.GetActor(f099addr)
+			incrementaddr = f099addr
+			if err != nil {
+				return cid.Undef, err
+			}
+			// if you don't find THAT one, you really messed up bad!
+			if !found {
+				return cid.Undef, xerrors.Errorf("could not find actor for the owner of the deleted miner, and then could not find f099 to send the funds to as a backup. something is very wrong here.")
+			}
+		}
+		actor.Balance = big.Add(actor.Balance, valueT)
+		err = actorsOut.SetActor(incrementaddr, actor)
+		if err != nil {
+			return cid.Undef, err
+		}
 	}
 
 	elapsed := time.Since(startTime)
@@ -232,9 +490,17 @@ type actorMigrationInput struct {
 	cache      MigrationCache  // cache of existing cid -> cid migrations for this actor
 }
 
+type balanceTransferInfo struct {
+	address address.Address
+	value   big.Int
+}
+
 type actorMigrationResult struct {
-	newCodeCID cid.Cid
-	newHead    cid.Cid
+	newCodeCID                            cid.Cid
+	newHead                               cid.Cid
+	minerTypeMigrationShouldDelete        bool
+	minerDeletedWasAboveMinPower          bool
+	minerTypeMigrationBalanceTransferInfo balanceTransferInfo
 }
 
 type actorMigration interface {
@@ -252,8 +518,11 @@ type migrationJob struct {
 }
 
 type migrationJobResult struct {
-	address.Address
-	states6.Actor
+	address                               address.Address
+	actor                                 states6.Actor
+	minerTypeMigrationShouldDelete        bool
+	minerDeletedWasAboveMinPower          bool
+	minerTypeMigrationBalanceTransferInfo balanceTransferInfo
 }
 
 func (job *migrationJob) run(ctx context.Context, store cbor.IpldStore, priorEpoch abi.ChainEpoch) (*migrationJobResult, error) {
@@ -270,9 +539,18 @@ func (job *migrationJob) run(ctx context.Context, store cbor.IpldStore, priorEpo
 	}
 
 	// Set up new actor record with the migrated state.
+	// XXX: to test: add one of each type of miner, maybe add some sectors, make a complex enough state and check some invariants???
+	// XXX: give some some fees, give some no fees, etc, etc, etc
+	// XXX: what happens if someone sends these addresses some funds??? no good.
+	// XXX: give some some fees, give some no fees, etc, etc, etc
+	// XXX: https://github.com/filecoin-project/specs-actors/blob/0fa32a654d910960306a0567d69f8d2ac1e66c67/actors/migration/nv4/top.go#L228
+	// XXX: https://github.com/filecoin-project/specs-actors/issues/1474#issuecomment-948948954
 	return &migrationJobResult{
-		job.Address, // Unchanged
-		states6.Actor{
+		minerTypeMigrationShouldDelete:        result.minerTypeMigrationShouldDelete,
+		minerDeletedWasAboveMinPower:          result.minerDeletedWasAboveMinPower,
+		minerTypeMigrationBalanceTransferInfo: result.minerTypeMigrationBalanceTransferInfo,
+		address:                               job.Address, // Unchanged
+		actor: states6.Actor{
 			Code:       result.newCodeCID,
 			Head:       result.newHead,
 			CallSeqNum: job.Actor.CallSeqNum, // Unchanged
@@ -288,8 +566,11 @@ type nilMigrator struct {
 
 func (n nilMigrator) migrateState(_ context.Context, _ cbor.IpldStore, in actorMigrationInput) (*actorMigrationResult, error) {
 	return &actorMigrationResult{
-		newCodeCID: n.OutCodeCID,
-		newHead:    in.head,
+		newCodeCID:                            n.OutCodeCID,
+		newHead:                               in.head,
+		minerTypeMigrationShouldDelete:        false,
+		minerDeletedWasAboveMinPower:          false,
+		minerTypeMigrationBalanceTransferInfo: balanceTransferInfo{},
 	}, nil
 }
 
@@ -314,11 +595,13 @@ func (c cachedMigrator) migrateState(ctx context.Context, store cbor.IpldStore, 
 		return nil, err
 	}
 	return &actorMigrationResult{
-		newCodeCID: c.migratedCodeCID(),
-		newHead:    newHead,
+		newCodeCID:                            c.migratedCodeCID(),
+		newHead:                               newHead,
+		minerTypeMigrationShouldDelete:        false,
+		minerDeletedWasAboveMinPower:          false,
+		minerTypeMigrationBalanceTransferInfo: balanceTransferInfo{},
 	}, nil
 }
-
 func cachedMigration(cache MigrationCache, m actorMigration) actorMigration {
 	return cachedMigrator{
 		actorMigration: m,
